+++ strawberry/src/pages/newCar/components/NewCarPageWrapper.tsx
@@ -1,0 +1,12 @@
+import { NewCarContextProvider } from "../contexts/NewCarContext";
+import NewCarPage from "../NewCarPage";
+
+function NewCarPageWrapper() {
+  return (
+    <NewCarContextProvider>
+      <NewCarPage />
+    </NewCarContextProvider>
+  );
+}
+
+export default NewCarPageWrapper;
