+++ strawberry/src/router/router.tsx
@@ -14,7 +14,7 @@ import DrawingLandingWrapper from "../pages/drawingLanding/DrawingLandingWrapper
 import DrawingPlayWrapper from "../pages/drawingPlay/DrawingPlayWrapper";
 
 import LandingPage from "../pages/landing/LandingPage";
-import NewCarPage from "../pages/newCar/components/NewCarPage";
+import NewCarPageWrapper from "../pages/newCar/components/NewCarPageWrapper";
 
 const router = createBrowserRouter([
   {
@@ -41,7 +41,7 @@ const router = createBrowserRouter([
         path: "introduce",
         element: (
           <>
-            <NewCarPage />
+            <NewCarPageWrapper />
           </>
         ),
       },
